pyproject.toml
[project]
name = "llm-image-analyzer-mcp"
version = "0.1.0"
description = "MCP server for analyzing images using Azure OpenAI vision models"
readme = "README.md"
authors = [
    { name = "Torstein Sørnes", email = "t.soernes@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=0.5.0",
    "pydantic-ai>=0.0.49",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
llm-image-analyzer-mcp = "llm_image_analyzer_mcp.server:main"

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
